Externalized snippet menu items
