+++ src/components/menus/SnippetList.jsx
@@ -1,48 +1,9 @@
 import React, { Component, PropTypes } from 'react';
-import Moment from 'react-moment';
 import { MenuItem } from 'material-ui';
 import ArrowDropRight from 'material-ui/svg-icons/navigation-arrow-drop-right';
-import { mapLanguage } from '../../util/rules';
 import CustomPropTypes from '../../util/custom-prop-types';
 
-const styles = {
-  menuItem: {
-    width: '350px',
-  },
-  title: {
-    width: '50%',
-  },
-  language: {
-    float: 'left',
-    fontSize: '80%',
-  },
-  lastEdited: {
-    float: 'right',
-    fontSize: '80%',
-  },
-};
-
-// Truncates the title if it is too long to fit in snippets dropdown
-const makeDisplayTitle = (title) => {
-  if (title.length > 10) {
-    return `${title.substring(0, 12)}...`;
-  }
-  return title;
-};
-
-const makeItemDiv = ({ snippetTitle, language, lastEdited }) => (
-  <table style={styles.menuItem}>
-    <tr>
-      <td style={styles.title}>{makeDisplayTitle(snippetTitle)}</td>
-      <td style={styles.language}>
-        {mapLanguage[language]}
-      </td>
-      <td style={styles.lastEdited}>
-        <Moment fromNow ago>{lastEdited}</Moment>
-      </td>
-    </tr>
-  </table>
-);
+import SnippetMenuItem from './SnippetMenuItem';
 
 const makeMenuItems = (snippetOwner, snippetsList, onClick) => {
   if (Object.keys(snippetsList).length === 0) {
@@ -52,7 +13,7 @@ const makeMenuItems = (snippetOwner, snippetsList, onClick) => {
     <MenuItem
       key={snippetKey}
       onTouchTap={() => onClick(snippetOwner, snippetKey)}
-      primaryText={makeItemDiv(snippetsList[snippetKey])}
+      primaryText={SnippetMenuItem(snippetsList[snippetKey])}
     />
   ));
 };
